+++ server/runtime_lua_nakama.go
@@ -2124,9 +2124,10 @@ func (n *RuntimeLuaNakamaModule) authenticateSteam(l *lua.LState) int {
 // @param userId(type=string) User ID to use to generate the token.
 // @param username(type=string, optional=true) The user's username. If left empty, one is generated.
 // @param expiresAt(type=number, optional=true) UTC time in seconds when the token must expire. Defaults to server configured expiry time.
+// @param refreshTokenExpiresAt(type=number, optional=true) UTC time in seconds when the token must expire. Defaults to server configured refresh token expiry time.
 // @param vars(type=table, optional=true) Extra information that will be bundled in the session token.
 // @return token(string) The Nakama session token.
-// @return validity(number) The period for which the token remains valid.
+// @return refreshToken(string) The Nakama session refresh token.
 // @return error(error) An optional error value if an error occurred.
 func (n *RuntimeLuaNakamaModule) authenticateTokenGenerate(l *lua.LState) int {
 	// Parse input User ID.
@@ -2147,13 +2148,19 @@ func (n *RuntimeLuaNakamaModule) authenticateTokenGenerate(l *lua.LState) int {
 		username = generateUsername()
 	}
 
-	exp := l.OptInt64(3, 0)
-	if exp == 0 {
-		// If expiry is 0 or not set, use standard configured expiry.
-		exp = time.Now().UTC().Add(time.Duration(n.config.GetSession().TokenExpirySec) * time.Second).Unix()
+	tokenExp := l.OptInt64(3, 0)
+	if tokenExp == 0 {
+		// If expiry is 0 or not set, use default configured token expiry.
+		tokenExp = time.Now().UTC().Add(time.Duration(n.config.GetSession().TokenExpirySec) * time.Second).Unix()
 	}
 
-	vars := l.OptTable(4, nil)
+	refreshTokenExp := l.OptInt64(4, 0)
+	if refreshTokenExp == 0 {
+		// If expiry is 0 or not set, use default configured refresh token expiry.
+		refreshTokenExp = time.Now().UTC().Add(time.Duration(n.config.GetSession().TokenExpirySec) * time.Second).Unix()
+	}
+
+	vars := l.OptTable(5, nil)
 	var varsMap map[string]string
 	if vars != nil {
 		var conversionError string
@@ -2176,16 +2183,18 @@ func (n *RuntimeLuaNakamaModule) authenticateTokenGenerate(l *lua.LState) int {
 		})
 
 		if conversionError != "" {
-			l.ArgError(4, conversionError)
+			l.ArgError(5, conversionError)
 			return 0
 		}
 	}
 
-	token, exp := generateTokenWithExpiry(n.config.GetSession().EncryptionKey, userIDString, username, varsMap, exp)
-	n.sessionCache.Add(uid, exp, token, 0, "")
+	tokenId := uuid.Must(uuid.NewV4()).String()
+	token, tokenExp := generateTokenWithExpiry(n.config.GetSession().EncryptionKey, tokenId, userIDString, username, varsMap, tokenExp)
+	refreshToken, refreshTokenExp := generateTokenWithExpiry(n.config.GetSession().EncryptionKey, tokenId, userIDString, username, varsMap, refreshTokenExp)
+	n.sessionCache.Add(uid, tokenExp, tokenId, refreshTokenExp, tokenId)
 
 	l.Push(lua.LString(token))
-	l.Push(lua.LNumber(exp))
+	l.Push(lua.LString(refreshToken))
 	return 2
 }